# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/te/eventEmitter": {
      "post": {
        "operationId": "TeController_pfEventEmitter",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/pfDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Torus API"
        ]
      }
    },
...
    "/tob_api_process_logs/schema": {
      "get": {
        "description": "Retrive the datatype of the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_findSchema",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_api_process_logsEntity"
                }
              }
            }
          }
        },
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_api_process_logs/get": {
      "get": {
        "description": "Filter all the records from the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_api_process_logsEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_api_process_logs/{id}": {
      "get": {
        "description": "Read all the records from the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_api_process_logsEntity"
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Updatetob_api_process_logsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_api_process_logsEntity"
                }
              }
            }
          }
        },
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_api_process_logsEntity"
                }
              }
            }
          }
        },
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_api_process_logs/query": {
      "post": {
        "description": "Read all the records from the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_findAllwithquery",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Querytob_api_process_logsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_api_process_logsEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/tob_api_process_logs": {
      "get": {
        "description": "Read all the records from the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_findAll",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/tob_api_process_logsEntity"
                  }
                }
              }
            }
          }
        },
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the tob_api_process_logs table",
        "operationId": "tob_api_process_logsController_create",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Createtob_api_process_logsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/tob_api_process_logsEntity"
                }
              }
            }
          }
        },
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    }
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://tgadev.gsstvl.com/ct242/tob001/tob002/v1/api-int",
      "description": "Production Server"
    }
  ],
  "components": {
    "schemas": {
...
      "pfDto": {
        "type": "object",
        "properties": {}
      },
...
      "tob_api_process_logsEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "apiendpoint": {
            "type": "string"
          },
          "requestdata": {
            "type": "object"
          },
          "responsedata": {
            "type": "object"
          },
          "apiname": {
            "type": "string"
          },
          "sample": {
            "type": "string"
          },
          "tob_consent_requestid": {
            "type": "string"
          },
          "tob_api_repositoryid": {
            "type": "string"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "apiendpoint",
          "requestdata",
          "responsedata",
          "apiname",
          "sample",
          "tob_consent_requestid",
          "tob_api_repositoryid",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
      "Querytob_api_process_logsDto": {
        "type": "object",
        "properties": {
          "apiendpoint": {
            "type": "string"
          },
          "requestdata": {
            "type": "object"
          },
          "responsedata": {
            "type": "object"
          },
          "apiname": {
            "type": "string"
          },
          "sample": {
            "type": "string"
          },
          "tob_consent_requestid": {
            "type": "string"
          },
          "tob_api_repositoryid": {
            "type": "string"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "apiendpoint",
          "requestdata",
          "responsedata",
          "apiname",
          "sample",
          "tob_consent_requestid",
          "tob_api_repositoryid",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
      "Createtob_api_process_logsDto": {
        "type": "object",
        "properties": {
          "apiendpoint": {
            "type": "string"
          },
          "requestdata": {
            "type": "object"
          },
          "responsedata": {
            "type": "object"
          },
          "apiname": {
            "type": "string"
          },
          "sample": {
            "type": "string"
          },
          "tob_consent_requestid": {
            "type": "string"
          },
          "tob_api_repositoryid": {
            "type": "string"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "apiendpoint",
          "requestdata",
          "responsedata",
          "apiname",
          "sample",
          "tob_consent_requestid",
          "tob_api_repositoryid",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      },
      "Updatetob_api_process_logsDto": {
        "type": "object",
        "properties": {
          "apiendpoint": {
            "type": "string"
          },
          "requestdata": {
            "type": "object"
          },
          "responsedata": {
            "type": "object"
          },
          "apiname": {
            "type": "string"
          },
          "sample": {
            "type": "string"
          },
          "tob_consent_requestid": {
            "type": "string"
          },
          "tob_api_repositoryid": {
            "type": "string"
          },
          "trs_created_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_created_by": {
            "type": "string"
          },
          "trs_modified_date": {
            "type": "string",
            "format": "date-time"
          },
          "trs_modified_by": {
            "type": "string"
          },
          "trs_status": {
            "type": "string"
          },
          "trs_next_status": {
            "type": "string"
          },
          "trs_process_id": {
            "type": "string"
          },
          "trs_access_profile": {
            "type": "string"
          },
          "trs_org_grp_code": {
            "type": "string"
          },
          "trs_org_code": {
            "type": "string"
          },
          "trs_role_grp_code": {
            "type": "string"
          },
          "trs_role_code": {
            "type": "string"
          },
          "trs_ps_grp_code": {
            "type": "string"
          },
          "trs_ps_code": {
            "type": "string"
          }
        },
        "required": [
          "apiendpoint",
          "requestdata",
          "responsedata",
          "apiname",
          "sample",
          "tob_consent_requestid",
          "tob_api_repositoryid",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code"
        ]
      }
    }
  }
}